# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.BaseOrderRequest:
    properties:
      name:
        type: string
      qty:
        type: integer
    required:
    - name
    - qty
    type: object
  model.ConfirmPaymentRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  model.CreateOrderRequest:
    properties:
      customer_email:
        type: string
      orders:
        items:
          $ref: '#/definitions/model.BaseOrderRequest'
        type: array
    required:
    - customer_email
    type: object
  model.CreateOrderResponse:
    properties:
      customer_email:
        type: string
      message:
        description: Orders        []BaseOrder `json:"orders"`
        type: string
      order_id:
        type: integer
      total_price:
        type: number
    type: object
...
  model.OrderResponse:
    properties:
      order:
        type: object
    type: object
...
  web.ErrJSONResponse:
    properties:
      error:
        $ref: '#/definitions/web.Error'
        type: object
      process_time:
        type: string
      status:
        type: string
      success:
        example: false
        type: boolean
    type: object
  web.Error:
    properties:
      message:
        type: string
    type: object
  web.JSONResponse:
    properties:
      data:
        type: object
        x-omitempty: true
      process_time:
        type: string
      status:
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:9000
info:
  contact:
    email: support.family-catering@example.com
    name: Family catering API Support
    url: http://www.family-catering.com/support
  description: Documentation for Family Catering API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Family Catering API
  version: "1.0"
paths:
...
  /order:
    post:
      consumes:
      - application/json
      description: Create a new order
      parameters:
      - default: Bearer <your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            allOf:
            - $ref: '#/definitions/web.JSONResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.OrderResponse'
                  - properties:
                      order:
                        $ref: '#/definitions/model.CreateOrderResponse'
                    type: object
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Create order
      tags:
      - order
  /order/confirm/:
    put:
      description: Confirm unpaid order based on given email
      parameters:
      - default: Bearer <your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: customer email
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.ConfirmPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/web.JSONResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Confirm order payment
      tags:
      - order
  /order/search:
    get:
      consumes:
      - application/json
      description: Search an order based on given query params
      parameters:
      - default: Bearer <your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: menu'ss name (if more than one name separated with comma without
          additional space)
        in: query
        name: menu-names
        type: string
      - description: customer's email (if more than one email separated with comma
          without additional space)
        in: query
        name: emails
        type: string
      - description: menu's name
        in: query
        name: exact-names
        type: string
      - description: the number of menus ordered per menu
        in: query
        name: qty
        type: string
      - description: maximum price of menu
        in: query
        name: max-price
        type: string
      - description: minimum price of menu
        in: query
        name: min-price
        type: string
      - description: status or ordered menu
        in: query
        name: status
        type: string
      - description: ordered menu start at given day
        in: query
        name: start-day
        type: string
      - description: ordered menu end at given day
        in: query
        name: end-day
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/web.JSONResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Search order
      tags:
      - order
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"